Minor fixes for odo and slides

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/SlidesVertical.java
@@ -58,7 +58,7 @@ public class SlidesVertical {
 
 
     public int getEncoders() {
-        return (slidesLeft.getCurrentPosition() + slidesRight.getCurrentPosition()) / 2;
+        return (int)((slidesLeft.getCurrentPosition() + slidesRight.getCurrentPosition()) / 2);
     }
 
     //simply moves up or down based on input from controller
